create game_team_team inherited class
