+++ test/test_helper.rb
@@ -13,3 +13,4 @@ require './lib/team'
 require './lib/game_team_game'
 require './lib/game_team_league'
 require './lib/game_team_season'
+require './lib/game_team_team'
